+++ internal/service/vpc/resource_vpc_peering_connection.go
@@ -7,6 +7,8 @@ import (
 	"strings"
 	"time"
 
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/customdiff"
+
 	"github.com/aiven/aiven-go-client"
 	"github.com/aiven/terraform-provider-aiven/internal/schemautil"
 
@@ -97,6 +99,40 @@ func ResourceVPCPeeringConnection() *schema.Resource {
 
 		Schema:             aivenVPCPeeringConnectionSchema,
 		DeprecationMessage: "Please use a cloud specific VPC peering connection resource",
+		CustomizeDiff:      customdiff.If(schemautil.ResourceShouldNotExist, vpcCustomDiffPeeringConnectionExists()),
+	}
+}
+
+// Deprecated
+func vpcCustomDiffPeeringConnectionExists() func(ctx context.Context, d *schema.ResourceDiff, m interface{}) error {
+	return func(ctx context.Context, d *schema.ResourceDiff, m interface{}) error {
+		client := m.(*aiven.Client)
+
+		projectName, vpcID, err := schemautil.SplitResourceID2(d.Get("vpc_id").(string))
+		if err != nil {
+			return err
+		}
+
+		peerCloudAccount := d.Get("peer_cloud_account").(string)
+		peerVPC := d.Get("peer_vpc").(string)
+
+		var region *string
+		peerRegion := d.Get("peer_region").(string)
+		if peerRegion != "" {
+			region = &peerRegion
+		}
+
+		pc, err := client.VPCPeeringConnections.GetVPCPeering(
+			projectName, vpcID, peerCloudAccount, peerVPC, region)
+		if err != nil && !aiven.IsNotFound(err) {
+			return err
+		}
+
+		if pc != nil {
+			return fmt.Errorf("vpc peering connection already exists and cannot be created")
+		}
+
+		return nil
 	}
 }